move detector to operations directory and refactor
